move this check task to SparkSQLSessionManager.
