+++ externals/kyuubi-spark-sql-engine/src/main/scala/org/apache/kyuubi/engine/spark/session/SparkSQLSessionManager.scala
@@ -17,10 +17,13 @@
 
 package org.apache.kyuubi.engine.spark.session
 
+import java.util.concurrent.TimeUnit
+
 import org.apache.hive.service.rpc.thrift.TProtocolVersion
 import org.apache.spark.sql.SparkSession
 
 import org.apache.kyuubi.KyuubiSQLException
+import org.apache.kyuubi.config.KyuubiConf
 import org.apache.kyuubi.engine.spark.operation.SparkSQLOperationManager
 import org.apache.kyuubi.session._
 
@@ -75,7 +78,9 @@ class SparkSQLSessionManager private (name: String, spark: SparkSession)
   override def closeSession(sessionHandle: SessionHandle): Unit = {
     _latestLogoutTime = System.currentTimeMillis()
     super.closeSession(sessionHandle)
-    operationManager.removeSparkSession(sessionHandle)
+    val session = operationManager.removeSparkSession(sessionHandle)
+    info(s"${session.sparkContext.sparkUser}'s session with $sessionHandle is closed," +
+      s"current opening sessions $getOpenSessionCount")
   }
 
   private def setModifiableConfig(spark: SparkSession, key: String, value: String): Unit = {
@@ -87,4 +92,26 @@ class SparkSQLSessionManager private (name: String, spark: SparkSession)
   }
 
   override protected def isServer: Boolean = false
+
+  override def start(): Unit = {
+    startTimeoutChecker()
+    super.start()
+  }
+
+  private def startTimeoutChecker(): Unit = {
+    val interval = conf.get(KyuubiConf.ENGINE_CHECK_INTERVAL)
+    val idleTimeout = conf.get(KyuubiConf.ENGINE_IDLE_TIMEOUT)
+    val checkTask = new Runnable {
+      override def run(): Unit = {
+        while (getOpenSessionCount > 0 ||
+          System.currentTimeMillis - latestLogoutTime < idleTimeout) {
+          TimeUnit.MILLISECONDS.sleep(interval)
+        }
+        info(s"Idled for more than $idleTimeout ms, terminating")
+        sys.exit(0)
+      }
+    }
+    submitBackgroundOperation(checkTask)
+  }
+
 }
